first test works on manually inserted data

+++ x_view_core/test/test_neighbor_search.h
@@ -26,7 +26,7 @@ void generatePointsOnLine(const uint64_t N, const uint64_t D,
                           x_view::MatrixXr* points);
 
 void testKNN(const x_view::MatrixXr& points, const x_view::Vector3r& query,
-             const std::vector<uint64_t>& expected_indices);
+             const std::vector<int>& expected_indices);
 
 }
